% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizzou_document.R
\name{todo_document}
\alias{todo_document}
\title{todo document}
\usage{
todo_document(fig_width = 8, fig_height = 6, css = NULL,
  keep_md = FALSE, self_contained = TRUE, ...)
}
\arguments{
\item{fig_width}{Width in inches of plot}

\item{fig_height}{Height in inches}

\item{css}{Optional}

\item{keep_md}{Defaults to FALSE}

\item{self_contained}{Defaults to TRUE}

\item{...}{Other args to output_format}
}
\value{
html rendered document
}
\description{
Render a nice-looking TO DO list in HTML
}
